Changed duration format
Modified CallbackEstimateStimDurationPush method to display the calculated duration as hh:mm:ss.
I wanted to keep the miliseconds, but I couldn't configure the seconds format to both show miliseconds and keep the ss format when there's a single digit number left to the point. So it rounds it to full seconds only.