autoSyncSquare
added a checkbox feature in FFF for automating the contrast of syncSquare based on the background luminosity.
If checked, the syncSquareLuminosity property (set in VStim) is changed and initializeBackground (VStim method) is run again to apply the change.

I added this feature inside FFF since it is a simple stimulation, so it makes a good start. Other stimulations, espacially those with colored displays, might require different approachs for contrasting. For this reason, I find it best to implement this feature per stimulation as needed.